backend/requirement.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
google-ads
python-dotenv
google-generativeai
requests
beautifulsoup4